Add basic call recording.
So far ut has done call asserting - you assert
the calls you expect in advance and ut asserts
them as they occur.

I’ve now added call recording - you ask ut to 
record calls to a method, and its up to you to assert they happened as expected later.

This should be useful if the order of calls is
non-deterministic