feat: Add optional support for applying `gofmt` to the generated mocks
I noticed in a different project that `gofumpt -w .` passes generated
files through `gofmt`. This includes the mocks generated by ut. As the
mocks generated by ut are not `gofmt`-formatted, the operations
`go generate ./...` and `gofumpt -w .` do not commute.

This commit adds an (optional) flag to the tool to indicate to the tool
that mock files should be passed through `gofmt` before being written
out.